:truck: rename ReactionEmoji to Emoji
add a dummy ReactionEmoji class who extends Emoji to keep compatibility with oldest version
